feat: Move placeholder screens to separate files and update iOS target
Moves the placeholder screens (Home, Search, Favorites, and Profile) from `MainScreen.dart` into their own dedicated files within the `lib/ui/screens/` directory.

Also, updates the iOS deployment target from 12.0 to 13.0.
