+++ lib/ui/main/main.dart
@@ -22,7 +22,7 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
       ),
-      home: const MainScreen(), // Changed to MainScreen
+      home: const MainScreen(),
     );
   }
 }
