+++ lib/ui/screens/HomeScreen.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+import 'package:sellio_mobile/core/design_system/themes/sellio_theme_provider.dart';
+
+class HomeScreen extends StatelessWidget {
+  const HomeScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text(
+        'Home Screen',
+        style: context.theme.typography.textTheme.headlineLarge.copyWith(
+          color: context.theme.colors.primary,
+        ),
+      ),
+    );
+  }
+}
